code cleanup before review
